chore: qol improvements for development (#5)
There were a couple of things I realized while creating the Go HW that
we can
implement here.

The Nix shell needs better permissions, so I changed it to make sure the
owner
of the actual database was the chai user. No more database shenanigans
when
using the Nix shell hook to create the local DB.

Also, we need to load the `.env` file properly. This brings in
`godotenv` for
that; I don't check for development env or not because the `.env` file
shouldn't
exist on the machine, and if it does, we have bigger problems.